mmq: the last q8-sized arithmetic on the E4M3 staging path, and the tile_C leftover
Types-sweep batch 4:
- ds4_quantize/gather_mmq_e4m3: the z-slice block index divided by QK8_1
  (the q8_1 block length) -- numerically identical to the correct
  DS4_ACT_BLOCK_VALS expression only because 4*32 == 128, and inert only
  because every live launch has gridDim.z == 1.  Now sized by the block
  the buffer actually holds; that was QK8_1's last consumer, so the
  define leaves ggml-common.h (iq2xxs_grid is all that remains).
- tile_C was tile<16,8,int>, a leftover of the deleted s8/s32 arm, safe
  by three coincidences (ne is T-independent; x[] never read; lane maps
  agree).  Now <float>, matching the FFMA accumulator that flows; A/B's
  'int' documented as four packed E4M3 bytes per register.
- the two dead 'blck = ggml_blck_size(type)' locals removed (last live
  callers of the stub's q8-block sizing).

Bit-exact by identity at every point.

+++ src/cuda/mmq/ds4_mmq_d2r.cu
@@ -923,9 +923,15 @@ void gateup_iq2_d2r_pair_kernel(const void * __restrict__ gate_soa,
         return;
     }
 
+    /* A/B: `int` here means "one 32-bit register holding four E4M3 bytes",
+     * not an integer element -- the packing is done by iq2_pack_e4m3_quad and
+     * load_B_tile.  C is genuinely f32 (the FFMA accumulator).  It was
+     * tile<16,8,int> -- a leftover of the deleted s8/s32 arm -- which was
+     * harmless only because tile::ne is T-independent and only ne/get_i/get_j
+     * are ever read; the stated type now matches what flows. */
     using tile_A = ggml_cuda_mma::tile<16, 8, int>;
     using tile_B = ggml_cuda_mma::tile<8, 8, int>;
-    using tile_C = ggml_cuda_mma::tile<16, 8, int>;
+    using tile_C = ggml_cuda_mma::tile<16, 8, float>;
 
     /* s_act / act_iter_base / act_stage etc. were s_q8 / q8_iter_base / q8_stage
      * until 2026-08-18.  The bytes have been E4M3 since the A8 campaign; only
